refactor: inline classes
